requirements.txt
canonicalwebteam.flask-base==0.7.3
canonicalwebteam.yaml-responses==1.1.1
canonicalwebteam.discourse==3.0.5
canonicalwebteam.search==0.2.1
canonicalwebteam.templatefinder==1.0.0
canonicalwebteam.image-template==1.3.1
black==20.8b1
flake8==3.8.4
